+++ packages/stream_chat/lib/src/core/api/responses.g.dart
@@ -461,10 +461,6 @@ RemoveUserGroupMembersResponse _$RemoveUserGroupMembersResponseFromJson(
   ..duration = json['duration'] as String?
   ..userGroup = UserGroup.fromJson(json['user_group'] as Map<String, dynamic>);
 
-SearchRolesResponse _$SearchRolesResponseFromJson(Map<String, dynamic> json) => SearchRolesResponse()
-  ..duration = json['duration'] as String?
-  ..roles = (json['roles'] as List<dynamic>?)?.map((e) => Role.fromJson(e as Map<String, dynamic>)).toList() ?? [];
-
 GetAppSettingsResponse _$GetAppSettingsResponseFromJson(
   Map<String, dynamic> json,
 ) => GetAppSettingsResponse()